Create distributed representation of documents

Create distributed representation of documents as weighted word vectors.

as.textmodel_doc2vec(
  x,
  model,
  normalize = FALSE,
  weights = 1,
  pattern = NULL,
  group_data = FALSE,
  ...
)

Arguments

x

a quanteda::tokens or quanteda::dfm object.

model

a textmodel_wordvector object.

normalize

if TRUE, normalized word vectors before creating document vectors.

weights

weight the word vectors by user-provided values; either a single value or multiple values sorted in the same order as the word vectors.

pattern

quanteda::pattern to select words to apply weights.

group_data

if TRUE, apply dfm_group(x) before creating document vectors.

...

additional arguments passed to quanteda::object2id.

Value

Returns a textmodel_docvector object with the following elements:

values

a list of matrices for word and document vectors.

dim

the size of the document vectors.

concatenator

the concatenator in x.

docvars

document variables copied from x.

normalize

if the document vectors are normalized.

call

the command used to execute the function.

version

the version of the wordvector package.